folder structure & renaming pages to views
