+++ src/components/Content/scss/content.scss
@@ -1,4 +1,4 @@
-@import "../../assets/scss/_vars";
+@import "../../../assets/scss/_vars";
 
 .content {
   position: relative;
