+++ src/components/Footer/scss/footer.scss
@@ -1,4 +1,4 @@
-@import "../../assets/scss/_vars";
+@import "../../../assets/scss/_vars";
 
 #footer {
   .social {
