+++ src/components/Header/scss/header.scss
@@ -1,4 +1,4 @@
-@import "../../assets/scss/_vars";
+@import "../../../assets/scss/_vars";
 
 #header {
   position: relative;
