+++ src/components/Overlay/scss/overlay.scss
@@ -1,4 +1,4 @@
-@import "../../assets/scss/_vars";
+@import "../../../assets/scss/_vars";
 
 .overlay {
   position: absolute;
