+++ src/components/Views/Contact.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Content from '../../Content/Content';
+import Content from '../Content/Content';
 
 class ContactComponent extends React.Component {
   render() {
