+++ src/components/Views/index.js
@@ -1,0 +1,7 @@
+import Home from './Home.js';
+import Contact from './Contact.js';
+
+export {
+  Home,
+  Contact
+};
